♻️ Move Description View Alert to UIComponents folder
